server/requirements.txt
flask>=2.3
flask-cors>=4.0
supabase>=2.4
python-dotenv>=1.0